requirements.txt
distribute==0.6.34
Django==1.5.1
MySQL-python==1.2.4
PIL==1.1.7
django-appconf==0.6
django-grappelli==2.4.5
django-imagekit==3.0.1
django-tinymce==1.5.1
pilkit==0.3.0
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
